chore: move nostr stuff into its own tab
