+++ fmo_frontend/src/components/mod.rs
@@ -5,6 +5,7 @@ mod copyable;
 mod federation;
 mod federations;
 mod navbar;
+pub mod nostr;
 mod tabs;
 
 pub use copyable::Copyable;
