+++ fmo_frontend/src/components/nostr/nostr_federation_row.rs
