+++ fmo_frontend/src/main.rs
@@ -1,3 +1,4 @@
+use fmo_frontend::components::nostr::NostrFederations;
 use fmo_frontend::components::{Federation, Federations, NavBar, NavItem};
 use leptos::*;
 use leptos_meta::{provide_meta_context, Link};
@@ -23,17 +24,19 @@ fn main() {
                             NavItem {
                                 name: "Home".to_owned(),
                                 href: "/".to_owned(),
-                                active: true,
+                                // TODO: make this actually work
+                                active: false,
                             },
                             NavItem {
-                                name: "About".to_owned(),
-                                href: "/about".to_owned(),
+                                name: "Nostr".to_owned(),
+                                href: "/nostr".to_owned(),
                                 active: false,
                             },
                         ]/>
                         <Routes>
                             <Route path="/" view=|| view! { <Federations/> }/>
                             <Route path="/federations/:id" view=|| view! { <Federation/> }/>
+                            <Route path="/nostr" view=|| view! { <NostrFederations/> }/>
                             <Route path="/about" view=|| view! { <div>About</div> }/>
                         </Routes>
                     </main>
